Add UserRestController OpenAPI definitions and fix permissions

+++ src/main/java/es/codeurjc/webapp15/controller/restController/UserRestController.java
@@ -3,11 +3,12 @@ package es.codeurjc.webapp15.controller.restController;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
-import es.codeurjc.webapp15.model.Artist;
 import es.codeurjc.webapp15.model.User;
 import es.codeurjc.webapp15.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.servlet.http.HttpServletRequest;
@@ -20,20 +21,17 @@ import java.net.URI;
 import java.security.Principal;
 import java.sql.SQLException;
 import java.util.Optional;
-import java.util.logging.Logger;
 
 import org.hibernate.engine.jdbc.BlobProxy;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.core.io.Resource;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -57,25 +55,65 @@ public class UserRestController {
     @Autowired
     private PasswordEncoder passwordEncoder;
 
+    @Operation (summary = "Gets the logged user")
+    @ApiResponses(value = {
+        @ApiResponse(
+            responseCode = "200",
+            description = "Found the user",
+            content = {@Content(
+                mediaType = "application/json",
+                schema = @Schema(implementation=User.class)
+                )}
+        ),
+        @ApiResponse(
+            responseCode = "401",
+            description = "User not authorized",
+            content = @Content
+        ),
+    })
     @GetMapping("/me")
     public ResponseEntity<Object> me(HttpServletRequest request) {
 
         Principal principal = request.getUserPrincipal();
-        if (principal != null) {
-            Optional<User> user = userService.findByEmail(principal.getName());
-            if (user.isPresent()) {
-                if (user.get().isRole("ADMIN") || user.get().isRole("USER")) {
-                    return ResponseEntity.ok(user);
-                }
-            }
+        if (principal == null) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
 
-        return ResponseEntity.notFound().build();
+        Optional<User> user = userService.findByEmail(principal.getName());
+        if (user.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        return ResponseEntity.ok(user);
     }
     
-
+    @Operation (summary = "Gets a user by its id")
+    @ApiResponses(value = {
+        @ApiResponse(
+            responseCode = "200",
+            description = "Found the user",
+            content = @Content
+        ),
+        @ApiResponse(
+            responseCode = "400",
+            description = "Invalid id supplied",
+            content = @Content
+        ),
+        @ApiResponse(
+            responseCode = "401",
+            description = "User not authorized",
+            content = @Content
+        ),
+        @ApiResponse(
+            responseCode = "404",
+            description = "User not found or doesn't have permission to access it",
+            content = @Content
+        ),
+    })
     @GetMapping("/{id}")
-    public ResponseEntity<Object> getUser(@PathVariable Long id, HttpServletRequest request) {
+    public ResponseEntity<Object> getUser(
+        @Parameter(description = "id of the user to be searched")
+        @PathVariable Long id, HttpServletRequest request) {
         
         Principal principal = request.getUserPrincipal();
         if (principal != null) {
@@ -103,14 +141,15 @@ public class UserRestController {
         ),
         @ApiResponse(
             responseCode = "400",
-            description = "Bad request, maybe one of the user attributes is missing or the type is not valid"
+            description = "Bad request, maybe one of the user attributes is missing or the type is not valid",
+            content = @Content
         ),
         @ApiResponse(
             responseCode = "409",
-            description = "User already exists"
+            description = "User already exists",
+            content = @Content
         )
     })
-
     @PostMapping("")
     public ResponseEntity<User> createUser(HttpServletRequest httpServletRequest, @RequestBody NewUser userBody){
         if((userBody.email == null) || (userBody.name == null) || (userBody.password == null)){
@@ -130,6 +169,29 @@ public class UserRestController {
         }
     }
 
+    @Operation (summary = "Gets the image of a user by its id")
+    @ApiResponses(value = {
+        @ApiResponse(
+            responseCode = "200",
+            description = "Found the user image",
+            content = @Content
+        ),
+        @ApiResponse(
+            responseCode = "400",
+            description = "Invalid id supplied",
+            content = @Content
+        ),
+        @ApiResponse(
+            responseCode = "401",
+            description = "User not authorized",
+            content = @Content
+        ),
+        @ApiResponse(
+            responseCode = "404",
+            description = "User not found, user image not found or doesn't have permission to access it",
+            content = @Content
+        ),
+    })
     @GetMapping("/{id}/image")
     public ResponseEntity<Object> getUserImage(@PathVariable Long id, HttpServletRequest request) throws SQLException {
 
@@ -161,8 +223,151 @@ public class UserRestController {
         return ResponseEntity.notFound().build();
     }
 
+    @Operation (summary = "Updates the image of a user")
+    @ApiResponses(value = {
+        @ApiResponse(
+            responseCode = "201",
+            description = "Image created correctly",
+            content = @Content
+        ),
+        @ApiResponse(
+            responseCode = "204",
+            description = "Image updated correctly",
+            content = @Content
+        ),
+        @ApiResponse(
+            responseCode = "400",
+            description = "Bad request",
+            content = @Content
+        ),
+        @ApiResponse(
+            responseCode = "401",
+            description = "User not authorized",
+            content = @Content
+        ),
+        @ApiResponse(
+            responseCode = "403",
+            description = "User not authorized",
+            content = @Content
+        ),
+        @ApiResponse(
+            responseCode = "404",
+            description = "User not found",
+            content = @Content
+        )
+    })
     @PutMapping("/{id}/image")
     public ResponseEntity<Object> postUserImage(@PathVariable Long id, @RequestParam("imageFile") MultipartFile file, HttpServletRequest request) throws IOException {
+
+        Principal principal = request.getUserPrincipal();
+        if (principal == null) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        Optional<User> user = userService.findByEmail(principal.getName());
+        if (user.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        if (!userService.exist(id)) {
+            return ResponseEntity.notFound().build();
+        }
+
+        if (user.get().isRole("ADMIN") ||
+            user.get().isRole("USER") && user.get().getId() == id) {
+
+                boolean hasImage = user.get().getImg_user() != null;
+
+                user.get().setImg_user(BlobProxy.generateProxy(file.getInputStream(), file.getSize()));
+                user.get().setImage(true);
+                userService.save(user.get());
+
+                URI location = fromCurrentRequest().buildAndExpand(user.get().getId()).toUri();
+
+                if (!hasImage) {
+                    return ResponseEntity.created(location).build();
+                }
+
+                return ResponseEntity.noContent().header("Location", location.toString()).build();
+        }
+        
+        return ResponseEntity.notFound().build();
+
+    }
+
+    @Operation(summary = "Update a user by its id")
+    @ApiResponses(value = {
+    @ApiResponse(
+    responseCode = "200",
+    description = "User updated correctly",
+    content = {@Content(
+    mediaType = "application/json",
+    schema = @Schema(implementation=User.class)
+    )}
+    ),
+    @ApiResponse(
+    responseCode = "400",
+    description = "User not updated",
+    content = @Content
+    ),
+    @ApiResponse(
+    responseCode = "403",
+    description = "User not authorized",
+    content = @Content
+    )
+    })
+    @PutMapping("/{id}")
+    public ResponseEntity<User> putUser(@PathVariable Long id, @RequestBody NewUser userBody) throws SQLException {
+
+        if (!userService.exist(id)) {
+            return ResponseEntity.notFound().build();
+        }
+
+        User storedUser = userService.findById(id).get();
+        String[] roles = storedUser.getRoles().toArray(new String[0]);
+
+        String encodedPassword = passwordEncoder.encode(userBody.password);
+        User updatedUser = new User(userBody.email, userBody.name, encodedPassword, roles);
+            
+        // Mantain current image
+        if (storedUser.getImg_user() != null) {
+            updatedUser.setImg_user(BlobProxy.generateProxy(storedUser.getImg_user().getBinaryStream(), storedUser.getImg_user().length()));
+        }
+
+        updatedUser.setId(id);
+        userService.save(updatedUser);
+
+        return ResponseEntity.ok(updatedUser);
+    }
+
+    @Operation(summary = "Delete a user by its id")
+    @ApiResponses(value = {
+    @ApiResponse(
+    responseCode = "200",
+    description = "User deleted correctly",
+    content = {@Content(
+    mediaType = "application/json",
+    schema = @Schema(implementation=User.class)
+    )}
+    ),
+    @ApiResponse(
+    responseCode = "400",
+    description = "User not deleted",
+    content = @Content
+    ),
+    @ApiResponse(
+    responseCode = "403",
+    description = "User not authorized",
+    content = @Content
+    ),
+    @ApiResponse(
+    responseCode = "404",
+    description = "User not found",
+    content = @Content
+    )
+    })
+    @DeleteMapping("/{id}")
+    public ResponseEntity<User> deleteUser(@PathVariable Long id, HttpServletRequest request) {
 
         Principal principal = request.getUserPrincipal();
         if (principal == null) {
@@ -178,92 +383,12 @@ public class UserRestController {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
 
-        if (user.get().isRole("ADMIN") ||
-            user.get().isRole("USER") && user.get().getId() == id) {
-
-                user.get().setImg_user(BlobProxy.generateProxy(file.getInputStream(), file.getSize()));
-                user.get().setImage(true);
-                userService.save(user.get());
-
-                URI location = fromCurrentRequest().path("/{id}/image").buildAndExpand(user.get().getId()).toUri();
-                return ResponseEntity.created(location).build();
+        Optional<User> requestedUser = userService.findById(id);
+        if (requestedUser.isEmpty()) {
+            return ResponseEntity.notFound().build();
         }
-        
-        return ResponseEntity.notFound().build();
 
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<User> putUser(@PathVariable Long id, @RequestBody User updatedUser) throws SQLException {
-
-        if (userService.exist(id)) {
-            
-            // Mantain current image
-            User storedUser = userService.findById(id).get();
-            if (storedUser.getImg_user() != null) {
-                updatedUser.setImg_user(BlobProxy.generateProxy(storedUser.getImg_user().getBinaryStream(), storedUser.getImg_user().length()));
-            }
-
-            updatedUser.setId(id);
-            userService.save(updatedUser);
-
-            return new ResponseEntity<>(updatedUser, HttpStatus.OK);
-
-        }
-        else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-    }
-    @PutMapping("/newname")
-    public ResponseEntity<User> changeInfo(HttpServletRequest request, @RequestBody String name) throws SQLException {
-        if(name==null){
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        }
-        Principal principal = request.getUserPrincipal();
-        if (principal != null) {
-            User user = userService.findByEmail(principal.getName()).get();
-            if((user.getName() != name)){
-                    user.setName(name);
-                    userService.save(user);
-                    return new ResponseEntity<>(user,HttpStatus.OK);
-            }
-        }
-        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-    }
-    @PutMapping("/newImage")
-    public ResponseEntity<User> changeInfo(HttpServletRequest request, @RequestBody MultipartFile image) throws SQLException {
-        if(image==null){
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        }
-        Principal principal = request.getUserPrincipal();
-        if (principal != null) {
-            User user = userService.findByEmail(principal.getName()).get();
-            if((user.getImg_user() != image)){
-                try {
-                    user.setImg_user(BlobProxy.generateProxy(image.getInputStream(), image.getSize()));
-                    user.setImage(true);
-                    userService.save(user);
-
-                    
-                    return new ResponseEntity<>(user,HttpStatus.OK);
-                }
-                catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
-        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Artist> deleteUser(@PathVariable Long id) {
-
-        try {
-            userService.delete(id);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-        }
-        catch (EmptyResultDataAccessException e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
+        userService.delete(id);
+        return ResponseEntity.ok(requestedUser.get());
     }
 }
